# backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessAuthDto"}}}}},"tags":["Auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessAuthDto"}}}}},"tags":["Auth"]}},"/auth/forgot-password":{"post":{"operationId":"AuthController_forgotPassword","parameters":[{"name":"email","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/verify-email":{"get":{"operationId":"AuthController_verifyEmail","parameters":[{"name":"code","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokensDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"tags":["Auth"]}},"/connections":{"post":{"operationId":"ConnectionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConnectionDto"}}}},"responses":{"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"tags":["Connections"]},"get":{"operationId":"ConnectionsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountDto"}}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"tags":["Connections"]}},"/connections/{id}":{"delete":{"operationId":"ConnectionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}},"tags":["Connections"]}},"/connections/twitter/oauth":{"get":{"operationId":"ConnectionsController_twitterAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["Connections"]}},"/connections/twitter/callback":{"get":{"operationId":"ConnectionsController_twitterCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["Connections"]}}},"info":{"title":"API","description":"API description","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AuthDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"user@example.com"},"password":{"type":"string","description":"Password of the user","example":"password"}},"required":["email","password"]},"TokensDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"accessTokenExpiresIn":{"format":"date-time","type":"string"},"refreshTokenExpiresIn":{"format":"date-time","type":"string"}},"required":["accessToken","refreshToken","accessTokenExpiresIn","refreshTokenExpiresIn"]},"UserDto":{"type":"object","properties":{"id":{"type":"number","description":"User ID"},"email":{"type":"string","description":"User email"},"isEmailVerified":{"type":"boolean","description":"User is email verified"}},"required":["id","email","isEmailVerified"]},"SuccessAuthDto":{"type":"object","properties":{"tokens":{"description":"Tokens","allOf":[{"$ref":"#/components/schemas/TokensDto"}]},"user":{"description":"User","allOf":[{"$ref":"#/components/schemas/UserDto"}]}},"required":["tokens","user"]},"ResetPasswordDto":{"type":"object","properties":{}},"ErrorResponse":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"CreateConnectionDto":{"type":"object","properties":{}},"AccountDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"username":{"type":"string"},"profilePicture":{"type":"string"},"socialNetworkId":{"type":"number","enum":[1,2,3],"items":{"type":"number","enum":["1","2","3"]}}},"required":["id","name","username","profilePicture","socialNetworkId"]}}}}